move specs next to their SUT rather than in a specific spec folder from root
