+++ .gitignore
@@ -1,4 +1,4 @@
-/node_modules
 .env
 /coverage
 /dist
+/node_modules
