+++ app/js/components/Tags.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {shallow} from 'enzyme'
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Tags from '../../app/js/components/Tags'
+import Tags from './Tags'
 
 Enzyme.configure({ adapter: new Adapter() });
