+++ app/js/components/uploader/Droparea.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {shallow} from 'enzyme'
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Droparea from '../../../app/js/components/uploader/Droparea'
+import Droparea from './Droparea'
 
 Enzyme.configure({ adapter: new Adapter() });
