+++ app/js/lib/useCases/PreviewUseCase.test.js
@@ -1,4 +1,4 @@
-import PreviewUseCase from '../../../app/js/lib/useCases/PreviewUseCase'
+import PreviewUseCase from './PreviewUseCase'
 import * as fs from 'fs';
 
 describe('PreviewUseCase', () => {
@@ -6,7 +6,7 @@ describe('PreviewUseCase', () => {
     // Arrange
     global.URL.createObjectURL = jest.fn();
     global.URL.createObjectURL.mockReturnValue('foo.jpeg')
-    const file = fs.createReadStream(__dirname + '/../../../app/images/test.jpg')
+    const file = fs.createReadStream(__dirname + '/../../../images/test.jpg')
     const previewUseCase = new PreviewUseCase([file])
 
     // Act
